Add bool-state testing
